accummulate fractional degree scrolling in pdf for scroll devices with higher resolution (mac os touchpad?)
git-svn-id: svn://svn.code.sf.net/p/texstudio/code/trunk@2834 92cce386-e96d-4106-809f-596c673a3fd8

+++ toolwidgets.cpp
@@ -89,8 +89,8 @@ void PreviewWidget::resetZoom(){
 void PreviewWidget::wheelEvent(QWheelEvent *event){
 	if (!preViewer->pixmap()) return;
 	if(event->modifiers()==Qt::ControlModifier){
-		int numDegrees = event->delta() / 8;
-		int numSteps = numDegrees / 15;
+		float numDegrees = event->delta() / 8.0f;
+		float numSteps = numDegrees / 15.0f;
 		scaleImage(pow(1.4,numSteps));
 		event->accept();
 	} else QScrollArea::wheelEvent(event);
